Renamed "bound_model"
