+++ packages/bindgen/src/bound-model.ts
